Refactored Resultintegrator to ChangeQueue
The datasource interface was slimmed down. The ChangeQueue is now owned
by the backend. The ChangeQueue does no longer keep track of which tasks
are currently being executed. Similarly, it is no longer possible to
wait for all pending tasks. The user has to manually keep reference to
tasks and check by itself.